# Revert "[SPARK-32550][SQL] Make SpecificInternalRow constructors faster by using while loops instead of maps