--- a/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/expressions/SpecificInternalRow.scala
+++ b/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/expressions/SpecificInternalRow.scala
@@ -206,7 +206,9 @@ final class SpecificInternalRow(val values: Array[MutableValue]) extends BaseGen
   }
 
   def this(dataTypes: Seq[DataType]) = {
-    // SPARK-32550: use while loop instead of map
+    // SPARK-32550: use `dataTypes.foreach` instead of `while loop + dataTypes(i)` to ensure
+    // constant-time access of dataTypes `Seq` because it is not necessarily an `IndexSeq` that
+    // support constant-time access.
     this(new Array[MutableValue](dataTypes.length))
     var i = 0
     dataTypes.foreach { dt =>