@LuciferYang 2020-08-24
@srowen you are right ~, I'll run a sql/catalyst module test and update it later