@LuciferYang 2020-08-24
@srowen @HyukjinKwon  Now the test performance returned to normal

```
Run completed in 7 seconds, 65 milliseconds.
Total number of tests run: 5
Suites: completed 2, aborted 0
Tests: succeeded 5, failed 0, canceled 0, ignored 0, pending 0
```