@HyukjinKwon 2020-08-24
Looks good to me too given the benchmark results.